requirements.txt
glfw==2.10.0
llvmlite==0.46.0
numba==0.63.1
numpy==2.3.5
PyOpenGL==3.1.10
PyOpenGL-accelerate==3.1.10
scipy==1.17.0